Reintroduced stepper motor support and unified profile values format

+++ src/ControlPlus/Port.h
@@ -25,8 +25,8 @@ private:
     {
         enum class Type { Value, Forward, Backward, Absolute, Unknown };
 
-        uint8_t value = 0;
-        uint8_t zeroValue = 0;
+        int value = 0;
+        int zeroValue = 0;
         Type type = Type::Unknown;
     };
